Remove unused URL-related methods and fix some minor typographical errors (#60)
